requirements.txt
openai
matplotlib
python-dotenv
langchain
langchain_openai
python-pptx
